web/client: created Ddi.RouteType
