+++ web/client/src/entities/Ddi/DdiProperties.tsx
@@ -20,7 +20,6 @@ export type DdiPropertyList<T> = {
     'conditionalRoute'?: T,
     'retailAccount'?: T,
     'target'?: T,
-    'companyFeatures'?: T,
 };
 
 export type DdiProperties = DdiPropertyList<Partial<PropertySpec>>;
