+++ web/client/src/entities/Ddi/Form.tsx
@@ -1,6 +1,5 @@
 import useFkChoices from '@irontec/ivoz-ui/entities/data/useFkChoices';
 import defaultEntityBehavior, { EntityFormProps, FieldsetGroups } from '@irontec/ivoz-ui/entities/DefaultEntityBehavior';
-import { PropertyList, ScalarProperty } from '@irontec/ivoz-ui/services/api/ParsedApiSpecInterface';
 import _ from '@irontec/ivoz-ui/services/translations/translate';
 import { useStoreState } from 'store';
 import { DdiPropertyList } from './DdiProperties';
@@ -9,7 +8,6 @@ import { foreignKeyGetter } from './foreignKeyGetter';
 const Form = (props: EntityFormProps): JSX.Element => {
 
     const { entityService, row, match } = props;
-    let properties = props.properties;
 
     const aboutMe = useStoreState((state) => state.clientSession.aboutMe.profile);
 
@@ -45,45 +43,6 @@ const Form = (props: EntityFormProps): JSX.Element => {
         match,
         skip,
     });
-
-    let overwriteProperties = false;
-    if (Object.keys(fkChoices).length) {
-
-        const companyFeatures = Object
-            .values(fkChoices.companyFeatures)
-            .map((row: any) => row.iden);
-
-        const routeType = {
-            ...properties.routeType,
-            enum: { ...(properties.routeType as ScalarProperty).enum },
-        };
-        const conditionalFeatures: Record<string, string> = {
-            'queues': 'queue',
-            'friends': 'friend',
-            'faxes': 'fax',
-            'conferences': 'conferenceRoom',
-        };
-        const conditionalFeaturesKeys = Object.keys(conditionalFeatures);
-
-        for (const conditionalFeature of conditionalFeaturesKeys) {
-
-            if (companyFeatures.includes(conditionalFeature)) {
-                continue;
-            }
-
-            delete routeType.enum[conditionalFeatures[conditionalFeature]];
-            overwriteProperties = true;
-        }
-
-        if (overwriteProperties) {
-            properties = {
-                ...props.properties,
-                routeType
-            };
-
-            entityService.replaceProperties(properties as PropertyList);
-        }
-    }
 
     const groups: Array<FieldsetGroups> = [
         {
